refactor: clean up dead code, add auth middleware, extract Chat.tsx hooks
Phase 1 — Dead code & shared utilities:
- Delete unused chat-new.ts (791 lines) and response-validator.ts (522 lines)
- Extract shared ensureDatabaseInitialized to database-init.ts
- Extract sanitizeErrorMessage to error-utils.ts

Phase 2 — Auth middleware:
- Add session-token auth (api/middleware/auth.ts)
- Login/register now return tokens; all protected routes require Bearer token
- Add fetchWithAuth helper; update 25 frontend fetch calls across 7 files

Phase 3 — Frontend refactor:
- Extract useConversations hook (conversation CRUD + storage sync)
- Extract useChatStream hook (SSE streaming + message state updates)
- Chat.tsx reduced from 1088 to 432 lines (-60%)
