+++ src/types/auth.ts
@@ -15,6 +15,7 @@ export interface User {
 
 export interface AuthState {
   user: User | null;
+  token: string | null;
   isAuthenticated: boolean;
   isLoading: boolean;
 }
